+++ Voxel2Pixel/Render/Array2xRenderer.cs
@@ -5,41 +5,41 @@
 	/// </summary>
 	public class Array2xRenderer : ArrayRenderer
 	{
-		public override void Tri(int x, int y, bool right, uint color)
+		public override void Tri(ushort x, ushort y, bool right, uint color)
 		{
 			if (right)
 			{
 				Rect(
-					x: x << 1,
+					x: (ushort)(x << 1),
 					y: y,
 					color: color,
 					sizeX: 2);
 				Rect(
-					x: x << 1,
-					y: y + 1,
+					x: (ushort)(x << 1),
+					y: (ushort)(y + 1),
 					color: color,
 					sizeX: 4);
 				Rect(
-					x: x << 1,
-					y: y + 2,
+					x: (ushort)(x << 1),
+					y: (ushort)(y + 2),
 					color: color,
 					sizeX: 2);
 			}
 			else
 			{
 				Rect(
-					x: (x + 1) << 1,
+					x: (ushort)((x + 1) << 1),
 					y: y,
 					color: color,
 					sizeX: 2);
 				Rect(
-					x: x << 1,
-					y: y + 1,
+					x: (ushort)(x << 1),
+					y: (ushort)(y + 1),
 					color: color,
 					sizeX: 4);
 				Rect(
-					x: (x + 1) << 1,
-					y: y + 2,
+					x: (ushort)((x + 1) << 1),
+					y: (ushort)(y + 2),
 					color: color,
 					sizeX: 2);
 			}
