+++ Voxel2Pixel/Render/ArrayRenderer.cs
@@ -8,7 +8,7 @@ namespace Voxel2Pixel.Render
 		public int Width { get; set; }
 		public int Height => (Image.Length / Width) >> 2;
 		#region IRectangleRenderer
-		public override void Rect(int x, int y, uint color, int sizeX = 1, int sizeY = 1) =>
+		public override void Rect(ushort x, ushort y, uint color, ushort sizeX = 1, ushort sizeY = 1) =>
 			Image.DrawRectangle(
 				x: x,
 				y: y,
